pubsub: rename awspubsub->awssnssqs and azurepubsub->azuresb
