+++ pubsub/awssnssqs/example_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package awspubsub_test
+package awssnssqs_test
 
 import (
 	"context"
@@ -23,7 +23,7 @@ import (
 	"github.com/aws/aws-sdk-go/service/sns"
 	"github.com/aws/aws-sdk-go/service/sqs"
 	"gocloud.dev/pubsub"
-	"gocloud.dev/pubsub/awspubsub"
+	"gocloud.dev/pubsub/awssnssqs"
 )
 
 func ExampleOpenTopic() {
@@ -42,7 +42,7 @@ func ExampleOpenTopic() {
 	// Construct a *pubsub.Topic.
 	// https://docs.aws.amazon.com/gettingstarted/latest/deploy/creating-an-sns-topic.html
 	topicARN := "arn:aws:service:region:accountid:resourceType/resourcePath"
-	t := awspubsub.OpenTopic(ctx, client, topicARN, nil)
+	t := awssnssqs.OpenTopic(ctx, client, topicARN, nil)
 	defer t.Shutdown(ctx)
 
 	// Now we can use t to send messages.
@@ -65,7 +65,7 @@ func ExampleOpenSubscription() {
 	// Construct a *pubsub.Subscription.
 	// https://docs.aws.amazon.com/sdk-for-net/v2/developer-guide/QueueURL.html
 	queueURL := "https://region-endpoint/queue/account-number/queue-name"
-	s := awspubsub.OpenSubscription(ctx, client, queueURL, nil)
+	s := awssnssqs.OpenSubscription(ctx, client, queueURL, nil)
 	defer s.Shutdown(ctx)
 
 	// Now we can use s to receive messages.
