+++ samples/gcmsg/main.go
@@ -28,8 +28,8 @@ import (
 
 	"github.com/google/subcommands"
 	"gocloud.dev/pubsub"
-	_ "gocloud.dev/pubsub/awspubsub"
-	_ "gocloud.dev/pubsub/azurepubsub"
+	_ "gocloud.dev/pubsub/awssnssqs"
+	_ "gocloud.dev/pubsub/azuresb"
 	_ "gocloud.dev/pubsub/gcppubsub"
 	_ "gocloud.dev/pubsub/rabbitpubsub"
 )
